update: 更新文件 common_context.cpp

+++ frameworks/js/napi/socket/async_context/src/common_context.cpp
@@ -49,7 +49,7 @@ int CommonContext::GetSocketFd() const
     if (sharedManager_ == nullptr) {
         return -1;
     }
-    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<intptr_t>(sharedManager_->GetData())) : -1;
+    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<uint64_t>(sharedManager_->GetData())) : -1;
 }
 
 bool CommonContext::CheckParamsType(napi_value *params, size_t paramsCount)
